// projects/gp-all-component/src/components/tables/gp-form/resources/gp-form-field-type.enum.ts
export enum GpFormFieldType {
  TEXT = 'gp-form-text-field',
  TEXT_AREA = 'gp-form-textarea-field',
  WYSIWYG = 'gp-form-wysiwyg-field',
  DROPDOWN = 'gp-form-dropdown-field',
  DROPDOWN_RELATED = 'gp-form-dropdown-related-field',
  IMG = 'gp-form-img-field',
  SWITCH = 'gp-form-switch-field',
  CHECKBOX = 'gp-form-checkbox-field',
  TIME = 'gp-form-time-field',
  CALENDAR = 'gp-form-calendar-field'
}